fixed typescript and added Matrix + Array tests
