+++ src/factoriesAny.js
@@ -220,9 +220,9 @@ export { createPinv } from './function/matrix/pinv.js'
 export { createEigs } from './function/matrix/eigs.js'
 export { createExpm } from './function/matrix/expm.js'
 export { createSqrtm } from './function/matrix/sqrtm.js'
-export { createSylvester } from './function/matrix/sylvester.js'
-export { createSchur } from './function/matrix/schur.js'
-export { createLyap } from './function/matrix/lyap.js'
+export { createSylvester } from './function/algebra/sylvester.js'
+export { createSchur } from './function/algebra/decomposition/schur.js'
+export { createLyap } from './function/algebra/lyap.js'
 export { createDivide } from './function/arithmetic/divide.js'
 export { createDistance } from './function/geometry/distance.js'
 export { createIntersect } from './function/geometry/intersect.js'
